refactored TestData to TestSupport, added constants, removed unused imports
